+++ packages/ts-newskit/src/components/slices/lead-story/__tests__/index.test.tsx
@@ -2,180 +2,64 @@ import React from 'react';
 import '@testing-library/jest-dom';
 import { render } from '../../../utils/test-utils';
 import { LeadStory } from '../index';
-import {
-  headline,
-  color,
-  readingTime,
-  summary,
-  bylines,
-  subHeadline,
-  caption,
-  image,
-  url
-} from '../../fixture/data.json';
+import { leadStory } from '../../../../slices/fixtures/data.json';
+
+const renderComponent = () => render(<LeadStory {...leadStory} />);
 
 describe('Render Component one', () => {
   it('should render a snapshot', () => {
-    const { asFragment } = render(
-      <LeadStory
-        headline={headline}
-        color={color}
-        readingTime={readingTime}
-        summary={summary}
-        bylines={bylines}
-        subHeadline={subHeadline}
-        caption={caption}
-        image={image}
-        url={url}
-      />
-    );
+    const { asFragment } = renderComponent();
     expect(asFragment()).toMatchSnapshot();
   });
 
   it('should render correct headline', () => {
-    const { getByText } = render(
-      <LeadStory
-        headline={headline}
-        color={color}
-        readingTime={readingTime}
-        summary={summary}
-        bylines={bylines}
-        subHeadline={subHeadline}
-        caption={caption}
-        image={image}
-        url={url}
-      />
-    );
-    const headlineText = getByText(headline);
+    const { getByText } = renderComponent();
+    const headlineText = getByText(leadStory.headline);
     expect(headlineText).toBeInTheDocument();
   });
   it('should render correct subHeadline color', () => {
-    const { getByText } = render(
-      <LeadStory
-        headline={headline}
-        color={color}
-        readingTime={readingTime}
-        summary={summary}
-        bylines={bylines}
-        subHeadline={subHeadline}
-        caption={caption}
-        image={image}
-        url={url}
-      />
-    );
+    const { getByText } = renderComponent();
 
     const subHeadlineText = getByText('JOEL KPOKU INTERVIEW');
     expect(subHeadlineText).toHaveStyle('color: #008347');
   });
   it('should render correct summary', () => {
-    const { getByText } = render(
-      <LeadStory
-        headline={headline}
-        color={color}
-        readingTime={readingTime}
-        summary={summary}
-        bylines={bylines}
-        subHeadline={subHeadline}
-        caption={caption}
-        image={image}
-        url={url}
-      />
-    );
+    const { getByText } = renderComponent();
 
-    const summaryText = getByText(summary);
+    const summaryText = getByText(leadStory.summary);
     expect(summaryText).toBeInTheDocument();
   });
   it('should render correct readingTime', () => {
-    const { getByText } = render(
-      <LeadStory
-        headline={headline}
-        color={color}
-        readingTime={readingTime}
-        summary={summary}
-        bylines={bylines}
-        subHeadline={subHeadline}
-        caption={caption}
-        image={image}
-        url={url}
-      />
-    );
+    const { getByText } = renderComponent();
 
-    const readingTimeText = getByText(readingTime);
+    const readingTimeText = getByText(leadStory.readingTime);
     expect(readingTimeText).toBeInTheDocument();
   });
   it('should render correct bylines', () => {
-    const { getByText } = render(
-      <LeadStory
-        headline={headline}
-        color={color}
-        readingTime={readingTime}
-        summary={summary}
-        bylines={bylines}
-        subHeadline={subHeadline}
-        caption={caption}
-        image={image}
-        url={url}
-      />
-    );
+    const { getByText } = renderComponent();
 
-    const bylinesText = getByText(bylines);
+    const bylinesText = getByText(leadStory.bylines);
     expect(bylinesText).toBeInTheDocument();
   });
   it('should render correct caption', () => {
-    const { getByText } = render(
-      <LeadStory
-        headline={headline}
-        color={color}
-        readingTime={readingTime}
-        summary={summary}
-        bylines={bylines}
-        subHeadline={subHeadline}
-        caption={caption}
-        image={image}
-        url={url}
-      />
-    );
+    const { getByText } = renderComponent();
 
-    const captionText = getByText(caption);
+    const captionText = getByText(leadStory.caption);
     expect(captionText).toBeInTheDocument();
   });
   it('should render correct summary href', () => {
-    const { getByText } = render(
-      <LeadStory
-        headline={headline}
-        color={color}
-        readingTime={readingTime}
-        summary={summary}
-        bylines={bylines}
-        subHeadline={subHeadline}
-        caption={caption}
-        image={image}
-        url={url}
-      />
-    );
+    const { getByText } = renderComponent();
 
-    expect(getByText(summary).closest('a')).toHaveAttribute(
+    expect(getByText(leadStory.summary).closest('a')).toHaveAttribute(
       'href',
       'https://www.thetimes.co.uk'
     );
   });
 
   it('should render correct headline href', () => {
-    const { getByText } = render(
-      <LeadStory
-        headline={headline}
-        color={color}
-        readingTime={readingTime}
-        summary={summary}
-        bylines={bylines}
-        subHeadline={subHeadline}
-        caption={caption}
-        image={image}
-        url={url}
-      />
-    );
+    const { getByText } = renderComponent();
 
-    expect(getByText(headline).closest('a')).toHaveAttribute(
+    expect(getByText(leadStory.headline).closest('a')).toHaveAttribute(
       'href',
       'https://www.thetimes.co.uk'
     );