+++ packages/ts-newskit/src/slices/content-bucket-1/index.tsx
@@ -20,7 +20,11 @@ import {
   ArticleListItem,
   ArticleListItemProps
 } from '../../components/slices/articleList';
-import { LeadStoryDivider, LeadStoryCell, CellNoMargin } from './styles';
+import {
+  LeadStoryDivider,
+  LeadStoryCell,
+  CellNoMargin
+} from '../shared-styles';
 import { JournalistStack } from './journalist-stack';
 
 export interface ContentBucket1Props {
@@ -95,6 +99,7 @@ export const ContentBucket1 = ({
                   <ArticleListItem
                     {...article}
                     hasTopBorder={articleTopBorder}
+                    hideImage={breakpointKey === 'lg'}
                   />
                   {articleBorder}
                 </React.Fragment>