+++ packages/article-list/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.16.6](https://github.com/newsuk/times-components/compare/@times-components/article-list@9.16.5...@times-components/article-list@9.16.6) (2023-04-18)
+
+**Note:** Version bump only for package @times-components/article-list
+
+
+
+
+
 ## [9.16.5](https://github.com/newsuk/times-components/compare/@times-components/article-list@9.16.4...@times-components/article-list@9.16.5) (2023-04-06)
 
 **Note:** Version bump only for package @times-components/article-list