+++ packages/image/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.14.23](https://github.com/newsuk/times-components/compare/@times-components/image@6.14.22...@times-components/image@6.14.23) (2023-04-18)
+
+**Note:** Version bump only for package @times-components/image
+
+
+
+
+
 ## [6.14.22](https://github.com/newsuk/times-components/compare/@times-components/image@6.14.21...@times-components/image@6.14.22) (2023-04-06)
 
 **Note:** Version bump only for package @times-components/image